Extract MultiLanguageTest
